asi_evolve/requirements.txt
openai>=1.0.0
pyyaml>=6.0
jinja2>=3.0
numpy>=1.20.0
faiss-cpu>=1.7.0
sentence-transformers>=2.2.0
# Optional experiment tracking
wandb>=0.15.0
